+++ demisto_sdk/commands/content_graph/interface/graph.py
@@ -31,7 +31,7 @@ class ContentGraphInterface(ABC):
         pass
 
     @abstractmethod
-    def import_graph(self, imported_path: Optional[Path] = None) -> None:
+    def import_graph(self, imported_path: Optional[Path] = None) -> dict:
         pass
 
     @abstractmethod